use font chars

+++ src/gpu_2d/text/shaders/text.vert
@@ -16,7 +16,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
 layout(location = 0) in ivec2 window_coordinates;
-layout(location = 1) in ivec2 texture_coordinates;
+layout(location = 1) in vec2 texture_coordinates;
 
 uniform mat4 matrix;
